New spec for checking commutativity of accrueInterests
We show for supply/withdraw/borrow/repay that calling accrueInterests
before has no effect as these function already do this.
For all functions we show that calling accrueInterests before or after
yields the same result, except for setFeeRecipient, for which it makes
a difference (new fee recipient gets the interest if they are accrued
afterwards).